rename context file
